Add new task to store
