+++ src/App.svelte
@@ -2,9 +2,7 @@
   import { taskStore } from './lib/store/taskStore';
   import TaskList from './lib/Task/List.svelte';
   let currentActiveIndex = 0;
-  $: {
-    console.log(tasks);
-  }
+  $: {  }
 
   /**
    * @param {{ detail: { task: { id: number; description: string; completed: boolean; parentId: number }; }; }} e
@@ -14,27 +12,6 @@
     const parentTask = oldTasks.find((task) => task.id === e.detail.task.parentId);
     const completeTasks = parentTask.tasks.filter((task) => task.completed === true).length;
     parentTask.progressLevel = (completeTasks / parentTask.tasks.length) * 100;
-    tasks = oldTasks;
-  }
-
-  function addTask(e) {
-    const oldTasks = [...tasks];
-
-    const { description, id } = e.detail;
-    const parentTaskIndex = tasks.findIndex((task) => task.id === id);
-
-    const parentTask = oldTasks[parentTaskIndex];
-
-    const newTask = {
-      id: Math.floor(Math.random() * 100),
-      description,
-      completed: false,
-      parentId: id,
-    };
-
-    parentTask.tasks.push(newTask);
-    parentTask.progressLevel = (parentTask.tasks.filter((task) => task.completed === true).length / parentTask.tasks.length) * 100;
-    oldTasks[parentTaskIndex] = parentTask;
     tasks = oldTasks;
   }
 
@@ -61,7 +38,6 @@
           active={currentActiveIndex === id}
           {id}
           on:toggleTask={toggleTask}
-          on:addTask={addTask}
         />
       {/each}
     </div>
